requirements.txt
streamlit>=1.36
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
joblib>=1.3
matplotlib>=3.7
seaborn>=0.12
geopy>=2.3
requests>=2.31
python-dotenv>=1.0
